DF-6370 warning fix
Former-commit-id: 958dce379abcca5c3d12fe9ef7f5c808e1fcd2cc

+++ Sources/Internal/Render/Highlevel/Landscape.cpp
@@ -773,7 +773,7 @@ void Landscape::AllocateGeometryDataNoInstancing()
     if (quadsInWidth == 0)
         quadsInWidth = 1;
 
-    quadsInWidthPow2 = HighestBitIndex(quadsInWidth);
+    quadsInWidthPow2 = uint32(HighestBitIndex(quadsInWidth));
 
     for (uint32 y = 0; y < quadsInWidth; ++y)
     {
